% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{ad_to_bs}
\alias{ad_to_bs}
\title{Convert Anno Domini (AD) date to Bikram Sambat (BS) date}
\usage{
ad_to_bs(ad_date)
}
\arguments{
\item{ad_date}{A character vector or Date vector representing AD dates.}
}
\value{
A character vector of the equivalent BS dates in "YYYY-MM-DD" format.
}
\description{
This function is fully vectorized for high performance on large datasets.
}
\examples{
ad_to_bs("2023-09-01")
ad_to_bs(as.Date("2023-09-01"))
ad_to_bs(c("2023-09-01", "2023-10-15", NA))

}
\seealso{
\code{\link{bs_to_ad}} for converting BS to AD dates
}
