% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_date.R
\name{as.Date.bs_date}
\alias{as.Date.bs_date}
\title{Convert bs_date to AD Date}
\usage{
\method{as.Date}{bs_date}(x, ...)
}
\arguments{
\item{x}{A \code{bs_date} object.}

\item{...}{Ignored.}
}
\value{
A \code{Date} vector.
}
\description{
Convert bs_date to AD Date
}
\examples{
d <- bs_date("2080-05-15")
as.Date(d)
}
