% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_bs_date_from_days}
\alias{find_bs_date_from_days}
\title{Find BS date from total days}
\usage{
find_bs_date_from_days(year, days)
}
\arguments{
\item{year}{Numeric BS year}

\item{days}{Total number of days}
}
\value{
List with bs_year, bs_month, and bs_day
}
\description{
Find BS date from total days
}
\keyword{internal}
