% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_passed_days_in_year}
\alias{find_passed_days_in_year}
\title{Find passed days in a BS year}
\usage{
find_passed_days_in_year(year, month, day)
}
\arguments{
\item{year}{Numeric BS year}

\item{month}{Numeric month (1-12)}

\item{day}{Numeric day}
}
\value{
Total number of days passed in the year
}
\description{
Find passed days in a BS year
}
\keyword{internal}
