% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_date.R
\name{is.bs_date}
\alias{is.bs_date}
\title{Test if an object is a bs_date}
\usage{
is.bs_date(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical.
}
\description{
Test if an object is a bs_date
}
\examples{
is.bs_date(bs_date("2080-05-15"))
is.bs_date("2080-05-15")
}
