% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{ad_to_bs}
\alias{ad_to_bs}
\title{Convert Anno Domini (AD) date to Bikram Sambat (BS) date}
\usage{
ad_to_bs(ad_date)
}
\arguments{
\item{ad_date}{A character string representing an AD date in "YYYY-MM-DD" or "YYYY/MM/DD" format,
or a Date object}
}
\value{
A character string representing the equivalent BS date in "YYYY-MM-DD" format
}
\description{
Convert Anno Domini (AD) date to Bikram Sambat (BS) date
}
\examples{
ad_to_bs("2023-09-01")
ad_to_bs("2023/09/01")
ad_to_bs(Sys.Date())

}
\seealso{
\code{\link{bs_to_ad}} for converting BS to AD dates
}
