% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{ad_to_bs}
\alias{ad_to_bs}
\title{Convert Anno Domini (AD) date to Bikram Sambat (BS) date}
\usage{
ad_to_bs(ad_date)
}
\arguments{
\item{ad_date}{A character string, Date object, or vector representing AD dates in "YYYY-MM-DD" or "YYYY/MM/DD" format}
}
\value{
A character string or vector of strings representing the equivalent BS dates in "YYYY-MM-DD" format
}
\description{
Convert Anno Domini (AD) date to Bikram Sambat (BS) date
}
\examples{
ad_to_bs("2023-09-01")
ad_to_bs("2023/09/01")
ad_to_bs(Sys.Date())
ad_to_bs(c("2023-09-01", "2023-10-15"))

# Works with data frames
df <- data.frame(date = as.Date(c("2023-01-01", "2023-06-15", "2023-12-31")))
df$bs_date <- ad_to_bs(df$date)

}
\seealso{
\code{\link{bs_to_ad}} for converting BS to AD dates
}
