% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{bs_to_ad}
\alias{bs_to_ad}
\title{Convert Bikram Sambat (BS) date to Anno Domini (AD) date}
\usage{
bs_to_ad(bs_date)
}
\arguments{
\item{bs_date}{A character string or vector of strings representing BS dates in "YYYY-MM-DD" or "YYYY/MM/DD" format}
}
\value{
A character string or vector of strings representing the equivalent AD dates in "YYYY-MM-DD" format
}
\description{
Convert Bikram Sambat (BS) date to Anno Domini (AD) date
}
\examples{
bs_to_ad("2080-05-15")
bs_to_ad("2080/05/15")
bs_to_ad(c("2080-05-15", "2080-06-20"))

}
\seealso{
\code{\link{ad_to_bs}} for converting AD to BS dates
}
