% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nepali_month_name}
\alias{get_nepali_month_name}
\title{Get Nepali month name}
\usage{
get_nepali_month_name(month_number, lang = "en")
}
\arguments{
\item{month_number}{Numeric month (1-12)}

\item{lang}{Language for month name ("en" for English, "ne" for Nepali)}
}
\value{
Character string with month name
}
\description{
Get Nepali month name
}
\examples{
get_nepali_month_name(1)  # "Baishakh"
get_nepali_month_name(1, "ne")  # "\u092c\u0948\u0936\u093e\u0916"
}
